partially contracted IfieldGenerator and DecimalGenerator. changed random values for Decimal from long to int

+++ src/generators/DecimalGenerator.java
@@ -3,7 +3,7 @@ package generators;
 import java.util.Random;
 
 /**
- * 
+ * Generates Decimal values 
  * @author Brandon Kindred
  *
  */
@@ -12,9 +12,12 @@ public class DecimalGenerator implements IFieldGenerator<Double> {
 	public DecimalGenerator() {
 	}
 
+	/**
+	 * Generate double value by appending 2 integers
+	 */
 	public Double generateValue() {
 		Random random = new Random();
-		return Double.parseDouble(random.nextLong() + "." + random.nextLong());
+		return Double.parseDouble(random.nextInt() + "." + random.nextInt());
 	}
 
 }
